requirements.txt
asgiref==3.9.1
Django==5.2.5
django-environ==0.12.0
django-extensions==3.2.0
django-redis==6.0.0
djangorestframework==3.16.1
geographiclib==2.1
geopy==2.4.1
iniconfig==2.1.0
numpy==2.3.2
packaging==25.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.14.1
redis==6.4.0
shapely==2.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
